pyproject.toml
[tool.poetry]
name = "onion"
version = "0.1.0"
description = ""
authors = ["Suhendi Andigo <suhendiandigo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^5.4.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest-asyncio-cooperative = "^0.17.1"
black = "^21.6b0"
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
